[Tests] Start adding missing tests
