+++ src/lib/components/artwork_grids/__tests__/__fixtures__/artwork.ts
@@ -1,0 +1,24 @@
+export default () => {
+  return {
+    node: {
+      __id: "artwork-talking-politics",
+      title: "TALKING POLITICS",
+      date: "2008",
+      sale_message: "Not for sale",
+      is_in_auction: false,
+      image: {
+        url: "artsy.net/image-url",
+        aspect_ratio: 1.27,
+      },
+      artists: [
+        {
+          name: "Jenny Holzer",
+        },
+      ],
+      partner: {
+        name: "Kukje Gallery",
+      },
+      href: "/artwork/jenny-holzer-talking-politics",
+    },
+  }
+}
